+++ packages/flutter_secure_storage/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Update integration tests based on upstream flutter_secure_storage v10.2.0 (12 test cases).
+* Update integration tests.
 
 ## 0.1.3