+++ packages/keyboard_detection/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Add 18 integration test cases.
+* Update integration tests.
 
 ## 0.1.0